# backend/src/api/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: backend-api
basePath: /api
tags:
- name: items
consumes:
- application/json
produces:
- application/json
paths:
  /items:
    get:
      tags:
      - items
      summary: Find all instances of the model matched by filter from the data source.
      operationId: find
      parameters:
      - name: q
        in: query
        description: 'Search query'
        type: string
      responses:
        '200':
          description: Request was successful
          schema:
            type: array
            items:
              $ref: '#/definitions/ItemResumeResponse'
        '400':
          description: Invalid response from meli
      x-swagger-router-controller: Items
  '/items/{id}':
    get:
      tags:
      - items
      summary: 'Find a model instance by {{id}} from the data source.'
      operationId: findById
      parameters:
      - name: id
        in: path
        description: Item id
        required: true
        type: string
      responses:
        '200':
          description: Request was successful
          schema:
            $ref: '#/definitions/ItemDetailsResponse'
        '400':
          description: Invalid ID supplied
      x-swagger-router-controller: Items
definitions:
  Author:
    type: object
    properties:
      name:
        type: string
      lastname:
        type: string
  Price:
    type: object
    properties:
      currency:
        type: string
      amount:
        type: number
      decimals:
        type: number
  ItemDetails:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      price:
        $ref: '#/definitions/Price'
      picture:
        type: string
      condition:
        type: string
      free_shipping:
        type: boolean
      sold_quantity:
        type: number
      description:
        type: string
  ItemDetailsResponse:
    type: object
    properties:
      author:
        $ref: '#/definitions/Author'
      item:
        $ref: '#/definitions/ItemDetails'
  ItemResume:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      price:
        $ref: '#/definitions/Price'
      picture:
        type: string
      condition:
        type: string
      free_shipping:
        type: boolean
  ItemResumeResponse:
    type: object
    properties:
      author:
        $ref: '#/definitions/Author'
      items:
        type: array
        items:
          $ref: '#/definitions/ItemResume'